fix: suppress signed/unsigned and unused-param compiler warnings
- FDNReverb.cpp: loop indices int → size_t; remove unused kPi constant
- DiffusionNetwork.cpp: loop indices int → size_t
- DiffusionStage.h: silence unused feedbackState parameter
- EarlyReflections.cpp: loop indices int → size_t
- PluginProcessor.cpp: fix include paths to explicit relative paths

+++ Source/DSP/DiffusionNetwork/DiffusionNetwork.cpp
@@ -2,7 +2,7 @@
 
 void DiffusionNetwork::prepare(double sampleRate) {
     double ratio = sampleRate / 44100.0;
-    for (int s = 0; s < NUM_STAGES; ++s) {
+    for (size_t s = 0; s < static_cast<size_t>(NUM_STAGES); ++s) {
         int dL = static_cast<int>(DELAYS_L[s] * ratio) + 1;
         int dR = static_cast<int>(DELAYS_R[s] * ratio) + 1;
         stages[s].prepare(sampleRate, dL, dR);
@@ -21,7 +21,7 @@ void DiffusionNetwork::setNumStages(int n) {
 }
 
 void DiffusionNetwork::processStereo(float& sampleL, float& sampleR) {
-    for (int s = 0; s < numActive; ++s)
+    for (size_t s = 0; s < static_cast<size_t>(numActive); ++s)
         stages[s].processStereo(sampleL, sampleR);
 }
